lib/uklock: Add SMP safe mutex
The commit adds the SMP safe mutex, currently the mutex supports
recursion by default.
In previous implementation mutex used both lock count and the owner of
the lock to determine if lock is acquired or not. Since the current
thread is running on only one CPU at a time only the owner field will
suffice.
Whenever thread tries to acquire the mutex, the function checks if the
lock belongs to the thread or not. if yes then it simply increment the
lock count. otherwise the mutex sleeps until the owner is set to NULL
i.e. other thread releases the lock.

Signed-off-by: Sairaj Kodilkar <skodilkar7@gmail.com>