lib/uksched: Add the function to get stack bottom
This adds the function to get the stack bottom of the current process
stack, this pointer is used in the implemenetation of the reader writer
lock.
The stack bottom stores the pointer to the thread structure hence there
is one one mapping between these two. So instead of using the struct
thread pointer we can use stack bottom pointer. Advantage of using this
pointer is that we have tweleve free bits available to use since stack
is PAGE_SIZE align.

Signed-off-by: Sairaj Kodilkar <skodilkar7@gmail.com>