lib/uklock: Add lock function for reader and writer
The commit adds the uk_rwlock_rlock and uk_rwlock_wlock function.
Current implementation is basic and starves the writers.
The implementation uses the rwlock variable as owner in case of writer
lock and read count in case of reader lock. the first five bits of these
lock is reserved as lock hence the read count will have the shift of 5.
In case of writer lock it stores the stack bottom pointer instead of
actual thread structure pointer. This is because each thread pointer is
stored at the bottom of the stack and since stack is align to the page
size we can have unused bits depending on the PAGE_SIZE.

Signed-off-by: Sairaj Kodilkar <skodilkar7@gmail.com>