lib/uklock/mutex.c,Config.uk: Make mutex SMP safe
We can safely acquire a lock on uniprocessor systems by disabling the
interrupts, but this does not work in SMP environment. Hence, to make the
mutex SMP safe, this commit uses atomic compare exchange.
The mutex also associates an owner field with the lock. This field allows the
thread to acquire the same lock recursively.

Checkpatch-Ignore: LONG_LINE
Signed-off-by: Sairaj Kodilkar <skodilkar7@gmail.com>