lib/uklock: Add macros to define IRQ safe lock functions
The commit adds a macro which takes the lock type and function as
arguments. The macro then defines a new function which is irq safe.
There are three macros:
`_LOCK_IRQF()`: Defines a new function which saves irq flag before
acquiring lock.
`_TRYLOCK_IRQF()`: Defines a new function which saves irq flag before
trying to acquire the flags. It restores the flag if the lock is not
acquired.
`_UNLOCK_IRQF()`: Defines a new function which restores the irq flag after
releasing the lock.

Checkpatch-Ignore: LONG_LINE
Signed-off-by: Sairaj Kodilkar <skodilkar7@gmail.com>